multimodal-ai-assistant/backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
pymupdf
chromadb
sentence-transformers
python-docx
pandas
openpyxl
python-pptx
sqlalchemy
python-dotenv
passlib
python-jose[cryptography]
email-validator
